feat: Add ASDF_RUST_IGNORE_COMPONENTS
\#\# Description
Currently there's no way to skip installation of the docs (as well as
other components).
While helpful in general, on systems with a small amount of storage
(e.g. 128/256 GB) those megabytes can be a dealbreaker.
For the most recent stable version (1.51.0) it takes around __40%__ out
of
total space for the installation (__418/1145 MB__).

\#\# Proposal
This PR introduces `ASDF_RUST_IGNORE_COMPONENTS` which is being passed
as
a value of option `--without` to the Rust `install` script. This option
takes comma-separated list of components to skip during installation
(e.g. `rust-docs,miri-preview` to skip installation of docs and MIRI
preview)

\#\# Results
It greatly helps to save storage space.
As a side effect, it drastically decreases installation time: on my
machine, doc building takes a handful of time: __20/23 minutes__ of
installation (downloading + actual installation) was spent __building
docs__ I won't need.